+++ resolwe_bio/processes/variant_calling/gatk_variant_refinement.py
@@ -32,12 +32,12 @@ class GatkGenotypeRefinement(Process):
     name = "GATK refine variants"
     category = "GATK"
     process_type = "data:variants:vcf:refinevariants"
-    version = "1.0.1"
+    version = "1.1.0"
     scheduling_class = SchedulingClass.BATCH
     requirements = {
         "expression-engine": "jinja",
         "executor": {
-            "docker": {"image": "public.ecr.aws/genialis/resolwebio/dnaseq:6.1.0"}
+            "docker": {"image": "public.ecr.aws/genialis/resolwebio/dnaseq:6.3.1"}
         },
         "resources": {
             "cores": 2,